Update class conflicts to have a storage of classes
